chore: dedupe pomodoroStore tests
Port unique edge cases from src/tests/pomodoroStore.test.ts to canonical
src/tests/stores/pomodoroStore.test.ts and remove duplicate file.

Edge cases added:
- setSettings: empty object, zero, negative values
- setTimeLeft: negative, MAX_SAFE_INTEGER, decimal values
- addTask: empty string, whitespace, special chars, unicode, long text
- toggleTask: empty string ID
- removeTask: empty list, empty string ID, verify remaining tasks
- UI state: empty toast message
- resetTimer: custom settings, tasks unaffected
- Data integrity: state immutability

All 61 tests pass; TypeScript compiles clean.
